Filters: Allow field level overrides for showCount, pass through expanded (#861)
v1.4.0 QA: showCount and expanded should be passed through from the "fields" config in Facets.

TEST=manual

On local test experience, set "expanded: false" at a field-level in config with "expanded: true" at the component-level config. See field's filter un-expanded and the rest of the fields in the component expanded. Set "showCount: false" at a field-level config. See the count is hidden for the field, but it still displays for all the other fields.