Add ability to configure 'target' behavior of rich text content. (#854)
This PR adds a 'targetConfig' parameter to the SDK's formatRichText function. This
optional parameter, when specified, sets the 'target' attribute for the different RTF link types: link, email, and phone.

Note that a UUID is used by the RichTextFormatter as plugin names. This is due to a couple of factors:

- All invocations of RichTextFormatter#format use the same instance of
  Markdown-it.
- If multiple plugins share the same name, Markdown-it's disable method will
  only disable the first plugin it encounters with that name.

To prevent the plugin added during one invocation from impacting another invocation, we must name the plugins uniquely and disable them at the end of the method call.

J=SPR-2134
TEST=auto, manual

Added unit tests for the updated formatRichText function. Manually verified:
- Passing a 'target' of '_blank' to formatRichText added the correct 'target'
  attribute to all of the link anchor tags.
- When I did not provide a 'targetConfig' to formatRichText, the link anchor tags had no 'target' 
   attribute.
- Ensured that I could configure the 'target' behavior of the different types of links differently.