inject library version from git describe --tags (#831)
This is done to remove a version release step where we must make a
commit to change the library version in the SDK constants.js.

Library version is set to the output of git describe --tags. This will give us how far (in commit #) from a prior release tag we are as well as the current hash, if we are not currently on a tag.

We also want to have canary builds be identified by the commit hash (and persist for a long
time). We want to use dashes in the library versions to make it easier
to navigate the CDN url.

Canary builds will only be pushed to s3 for the develop branch.
As a result, /dev/branch CDN links will not be made for the develop branch.

These library version changes will affect three places we use the library:
1. In a release tag, used by the CDN
2. In a dist-* release tag
3. In the NPM registry

For 1 and 2, we rely on the gulp build process, which has been changed to
inject this build version from the current git repository.

For 3, we do not currently have the sdk in the npm registry. When we do,
this build can be included in a preinstall and inject correctly into the
files before pushing to the registry.

J=SPR-2346
TEST=manual

For 1: Tested to see that the circleci build for the PR has the git described tag in the answers.js (and propagated in the apirequest).

For 2: Tested locally on a dist tag created locally and deleted to
see a v# lib version number.

For 3: Tested the git commands for a test npm package on the registry on the
pre-install and it successfully got the tag value when the test package
was referenced in another repository.