Fix malformed calc in Card.scss (#818)
Division of the css variables performed outside of the calc in Card.scss
for the margin. This caused the FAQ to not have spacing in between
cards.

J=None
TEST=manual

Tested on a new Jambo site referencing a local sdk dist (served through npm
serve package). Verified that the FAQ cards did not have spacing before
and have margin between them now.