VerticalResults: only update when search complete
> Current: Removing an applied filter hides the entire component and causes a flash
Expected: No flash, the result count edit: and vertical results should remain
I think this is similar to the problem we have with pagination.
can we make it such that we wait to update the vertical results until the new set returns (instead of hiding it)?

TEST=manual
check that search results still update when removing a filter, applying a filter, and
making a new search