VerticalResults: use <h2> for yxt-Results-title
To match resultssectionheader.hbs

TEST=manual
check that answers.yext.com's yxt-Results-title has margin: 0 from the css reset.
We can expect all consulting sites to apply this css reset, so adding margin: 0 is not a breaking change.
check that visually looks the same out of the box whether it is h2 or div