VerticalResults: rename yxt-Results-viewMore -> yxt-Results-viewAll
In order to be more backwards compatible, we should use the older
viewAllLink css class. Styling for AccordionResults was not changed
with this commit, even though the yxt-Results-viewAllLink class is
now defined in two different places, because the preexisting AccordionResults
specific styling (luckily) completely overrides the new styling.

TEST=manual
checked in chrome inspector that the viewAllLink in AccordionResults
does not carry over any of the new styling. checked that for both on
hover and not, that all css properties from yxt-Results-viewAllLink
are overriden by the more specific .yxt-Accordion .yxt-Results-viewAllLink.