UniversalResults: make viewMore config option work (#879)
This option is supposed to prevent the viewMore div from
rendering if false. It was being ignored.

TEST=manual
check that I can set viewMore to false (and that the default is true))
and that the viewMore box will not render in this case. check that
setting it to true will render the box