VerticalResults: use ::after for viewAllLink icon
This commit changes the viewAllLink icon to use ::after styling instead of
an icon component, setting the desired chevron icon as a base64 encoded
background-image.

Generally, it is preferable to use an IconComponent when an icon is desired.
This mixin is for cases when we want to add an icon in a place where
SDK implementors commonly would add an icon with ::after. Using an IconComponent
in those cases would result in a double icon, while adding the desired icon with ::after
will only result in the new SDK icon being overriden.

TEST=manual
check that we still get a chevron icon with the ::after styling