Facets: fix expand field override
First fix was expand always being set to true if the top level was set to
true even if the fieldOverride expand was set to false.

Second fix was first checking the using the FilterOptions specific expand
over the FilterBox expand. Other attributes rely on Object.assign to correctly
be overridden, however going from FilterBox to FilterOptions the config name
changes from expand to showExpand so this doesn't work in this case.

It's super dumb that showExpand and expand are the same config option
but have different names.

TEST=manual
check that after the changes I can set
            fields: {
              c_employeeDepartment: {
                control: 'singleoption',
                expand: false
              }
            }
in my facets config, and the the Employee Department FilterOptions will
not be expandable