Facets: resetFacets undefined by default
FilterBox allows you to show a reset all button regardless of searchOnChange
as of #839. Facets also needed to be changed for this change, since it's a
wrapper over FilterBox. By default the resetButton should appear when
searchOnChange is false, and not be visible when searchOnChange is true.

TEST=manual
Test that in Facets,
with resetFacets unset
- searchOnChange: true hides the reset button
- searchOnChange: unset shows the reset button
- searchOnChange: false shows the reset button

---
- searchOnChange: true, resetFacets: true shows the reset button
- searchOnChange: false, resetFacets: false hides the reset button