Add a hardcoded default for js lib version when not in a git repository (#887)
* Add a hardcoded default for js lib version when not in a git repository

This is for the case when we are building the repository but without git
tracking. This is happening in the case of the netlify tests in our PRs
to master. Add this hardcoded default when we are not in a git working
tree.

J=None
TEST=manual

Check that after a gulp build, the three following files have (1) the
git describe --tags output as the js lib version as normal and (2) TEST
as the js lib version when the .git folder is removed.

* dist/answers.js
* dist/answers-umd.js
* dist/answers-modern.js

Verify the PR tests (starting with master to start the netlify tests,
then changing to correct base branch).