DirectAnswer: add cardOverrides (#865)
This commit adds cardOverrides to direct answers, which let you specify an
array of rules for specifying a specific custom component based on the
direct answer received. Currently supports matching by fieldName, fieldType,
and entityType.

defaultConfigOption() in configutils.js was renamed to getOrDefault() so that it could be used within this context

TEST=manual
tested with cardOverrides: [
{
  cardType: 'CustomDA1',
  fieldName: 'Phone Number'
},
{
  cardType: 'bogus component',
  fieldName: 'Phone Number'
}
]

checked that would use custom card for a phone number direct answer, but not for
others. checked that the first match would be used. checked that page did not
error when cardOverrides were commented out.