DirectAnswers: color: inherit for icon
.yxt-Results-titleIconWrapper was previously changed to have color: $color-brand-primary.
This css class was also being used in DirectAnswers, which caused an undesired color change.
Double checked, and yxt-Results-titleIconWrapper is only being used in DirectAnswers and
VerticalResults.

This commit adds the currently unused yxt-DirectAnswer-titleIconWrapper class to the
<div>. It has the exact same styling as yxt-Results-titleIconWrapper, except for color: inherit.
It's placed before yxt-Results-titleIconWrapper so that if somebody had custom css on
yxt-Results-titleIconWrapper, their custom css will have higher priority.

TEST=manual
test that by default the direct answers icon color is white