ResultsHeader: add yxt-Results-filters css class for compatibility (#903)
This commit adds the yxt-Results-filters css class to yxt-ResultsHeader,
so that if yxt-Results-filters was being display: none'd, then the ResultsHeader
will also be display: none'd. This is only done on universal. On universal, since
v0.13.0 the newer yxt-ResultsHeader has already been in use, and adding this class
on vertical pages is more likely to confuse us than help us in the future.

TEST=manual
test that on local verizon site with display: none, nlp filters no longer show
up on universal