VerticalResults: add query to view all link (#901)
If the url was passed in from the config, the `query` url param was not being set.
This is needed when going from universal to vertical search so that the same query is run on page load.

TEST=manual
test that locally, query will be passed through
test setting url to unset, null, undefined, '', defaults to <verticalKey>.html