ResultsHeader: display facets from response, not request
This commit moves applied filter display logic into the ResultsHeader,
instead of VerticalResults, and also sets up the component to listen
to changes to FACET_FILTER_NODE. Now, when the Facets component makes updates
to the stored facets, regardless of whether a new search is made or not,
the applied filters bar will update.

Originally, I wanted to just make VerticalResults listen to changes to
FACET_FILTER_NODE. However, VerticalResults already listens to changes
to VERTICAL_RESULTS, and the sdk does not play nice when a component
tries to listen to multiple StorageKeys.

A future commit will add collision handling when an NLP filter and facet
share the same display value + fieldId.

TEST=manual
test that results header will not render if showResultCount and appliedFilters.show are both false
test that facets/filters can still be removed from the applied filters bar
test that checking a facet, applying it, then removing it and applying again does not display the removed facet
  in the applied filters bar (i.e. we display facets received from the response, not sent in the request)
test that when I search "consulting", both an nlp filter tag and a facet removable filter tag will be rendered, and that
  clicking the removable tag will remove the selected facet and trigger a new search