Add context as URL parameter for Search API, add ANSWERS.setContext
We add the URL parameter for query rules in the Answers backend. A user
can set the context with the function and expect the context to persist
for subsequent searches in the same page (on both Universal and
Vertical).

Example:
ANSWERS.setContext({
  state: 'tx',
});

TEST=manual

Tested on a newly created local environment (using Jambo), testing a
local dist build of answers.

Make sure the context is not added by default by looking at the Network
tab in Chrome dev tools.

After calling ANSWERS.setContext (In both JavaScript and from Chrome console),
In both Universal and Vertical Results pages
Test query API requests of subsequent searches include context as a URL param
Test query API requests of application of a facet includes context as a URL parameter.
Test query API requests of pagination includes context as a URL parameter.