ResultsHeader: display facets from response, not request (#910)
This commit moves applied filter display logic into the ResultsHeader,
instead of VerticalResults, and also sets up the component to listen
to changes to DYNAMIC_FILTERS. Now, the applied filters bar will display the facets calculated from the response, instead of the facets sent in the request.

Originally, I wanted to just make VerticalResults listen to changes to
DYNAMIC_FILTERS. However, VerticalResults already listens to changes
to VERTICAL_RESULTS, and the sdk does not play nice when a component
tries to listen to multiple StorageKeys. In order to do this the filter calculation
logic needed to be moved into the ResultsHeader.

Due to some complications below, it was necessary to create a bespoke storageKey that mirrors DYNAMIC_FILTERS.

Just registering a listener to DYNAMIC_FILTERS inside the ResultsHeader doesn't work, though
because every time VerticalResults (the parent component) has its setState called, a new ResultsHeader component will be created from scratch, and that new component will register a new, duplicate event listener, without cleaning up the old one. Cleaning up the old event listener to DYNAMIC_FILTERS breaks Facets, which also listen to DYNAMIC_FILTERS, because the current implementation of the off() method erases all event listeners for a moduleId, and cannot be targeted at a specific listener. This problem was worked around by adding a unique ResultsHeader StorageKey, and SPR-2455 was created to investigate this spaghetti code.

A future commit will add collision handling when an NLP filter and facet
share the same display value + fieldId.

TEST=manual
test that results header will not render if showResultCount and appliedFilters.show are both false
test that facets/filters can still be removed from the applied filters bar
test that checking a facet, applying it, then removing it and applying again does not display the removed facet
  in the applied filters bar (i.e. we display facets received from the response, not sent in the request)
test that when I search "consulting", both an nlp filter tag and a facet removable filter tag will be rendered, and that
  clicking the removable tag will remove the selected facet and trigger a new search