Add referrerPageUrl to the query endpoint requests and persistentstorage
We add the referrerPageUrl as part of the query endpoint requests to
support query rules in the Answers backend. The URL of the current page
is the source of truth for this parameter. We make three changes to
support this:

* On ANSWERS.init, look at the current URL in persistent storage to get
  the initial value of referrerPageUrl.
* If there is no referrerPageUrl query param already, source it from
  document.referrer of the current page.
* Change the view more link to include referrerPageUrl in the
  navigation.

Note: This makes it such that referrerPageUrl is a query parameter that
is always on the page. This is because referrerPageUrl is determined
when the page loads and is added as a query parameter to maintain the
parameter as you navigate within the experience.

J=SPR-2452
TEST=manual

Test entry of page with URL param referrerPageUrl already set.

Test all search requests contain the referrerPageUrl in the request.

Test works when navigating to an answers page from a standalone page.
Add link to a standalone page to an answers root page. Click the link. Make sure
referrerPageUrl is the original standalone page.

Test as you navigate through the experience, the referrerPageUrl is
carried
 * Change filters
 * Navigation tabs
 * Alternative verticals
 * View more button

Test works in an iFrame, as you navigate within an iFrame, the
referrerPageUrl persists.