bust cache for xhr requests in css vars ponyfill for ie11 (#872)
* bust cache for xhr requests in css vars ponyfill for ie11

This change is necessary for the following situation:

Link tags make requests to servers without CORS headers. Because v1.3.1
(and all patch version CDN links) have a cache-control with a non-zero
max-age, this hints to a browser (especially IE11) that the requests
should be cached. Thus IE11, in the XMLHttpRequests made by the
css-vars-ponyfill, tries to access the request, is told it should used
the cache request, and finds that the cached request does not have CORS
headers and fails the XHRs.

We cannot make the CORS headers show up for the link tag without
removing the cache-control max-age or adding crossorigin attribute
(which, lucky for us, is not available in IE11). So, we must try to bust
the cache in IE11 for the XHRs made by css vars ponyfill.

This is done by adding a random string (current time) as the query
parameter to the links before they're sent. This _ query parameter is
used by jQuery to do similar cache busting.

We add a URI library that will allow us to add this query parameter
without worrying about the many cases of URL query parameter
construction.

T=330545
TEST=manual

Tested that all link tags in the affected Virtusa site (and a local
Jambo site) are fetched a second time with the cache busting query
param.

Tested a script with and without a query param already in the url.

With the link tag before the ANSWERS.init call, tested in IE11 (the
affected browser) and Chrome. Verified that the styling with css
variables are showing on IE11 on page load.

* Don't use urijs package, instead us URL consructor, polyfilled by core-js

* Add more documentation

* Surround in a try-catch, do not block js if URL is relative