DirectAnswer: add readme for custom DA card, fix fieldType
This commit adds a readme section for how to add a custom DA
card to the SDK. It also fixes the way fieldType is accessed when
calculating which cardOverride to use. One out of the 2 places
it is mentioned in the SDK spec is wrong and I only read the incorrect
one, and fieldType was not part of the query responses  at the time this
code was originally written.

TEST=manual
test that I can specify a cardOverride for fieldType: 'phone',
that will be applied to 'amani farooque phone' but not
'amani farooque employee department'