Readme: update radius: 0 readme explanation.
Addresses https://github.com/yext/answers/issues/915.
The readme was not being entirely truthful with what happens
When a radius filter has a value of 0 selected.
In this case no locationRadius param is passed to the backend,
but the backend is still free to do whatever custom thing they want.

TEST=none