move from search on context change to all URL components on context change
Update the individual components instead of running another search when
we update the context in global storage. This is because we anticipate
the setContext could be called on every ANSWERS.onReady. However, we
must continue passing the context in URLs for iFrame implementations.

Note: Because some of our URL generation functions calculate URLs ad-hoc
based on the URL at time of function call, and because
persistentStorage.set is behind a timeout, the functions may be called
before the URL reflects the updated context. Thus, we must pass in the
context from globalStorage instead of from the URL. In the future, it
would be good to have knowledge of when the URL was updated from the
persistentStorage.set (a callback system?) or to remove the gathering of
URL params from the current URL and instead rely on storage of these
params.