Add context as URL parameter for Search API, add ANSWERS.setContext (#916)
* Add context as URL parameter for Search API, add ANSWERS.setContext

We add the URL parameter for query rules in the Answers backend. A user
can set the context with the function and expect the context to persist
for subsequent searches in the same page (on both Universal and
Vertical).

Example:
ANSWERS.setContext({
  state: 'tx',
});

TEST=manual

Tested on a newly created local environment (using Jambo), testing a
local dist build of answers.

Make sure the context is not added by default by looking at the Network
tab in Chrome dev tools.

After calling ANSWERS.setContext (In both JavaScript and from Chrome console),
In both Universal and Vertical Results pages
Test query API requests of subsequent searches include context as a URL param
Test query API requests of application of a facet includes context as a URL parameter.
Test query API requests of pagination includes context as a URL parameter.

* move from search on context change to all URL components on context change

Update the individual components instead of running another search when
we update the context in global storage. This is because we anticipate
the setContext could be called on every ANSWERS.onReady. However, we
must continue passing the context in URLs for iFrame implementations.

Note: Because some of our URL generation functions calculate URLs ad-hoc
based on the URL at time of function call, and because
persistentStorage.set is behind a timeout, the functions may be called
before the URL reflects the updated context. Thus, we must pass in the
context from globalStorage instead of from the URL. In the future, it
would be good to have knowledge of when the URL was updated from the
persistentStorage.set (a callback system?) or to remove the gathering of
URL params from the current URL and instead rely on storage of these
params.

* Use globalStorage instead of persistentStorage

* Add validation for context

* Source from globalStorage from getState, as they are different

* Updates to validation

- Use allowlist for typing validation
- Do not set a default param value (for case of setContext(undefined))
- Stringify once
- Add console.error for setContext

* update size limit to 155kb for context change

* Use addParamsToUrl instead of SearchParams + '?'

* Styling changes for function calls