Create Percy snapshots when running acceptance tests.
This PR adds Percy to the SDK's testing toolchain. Now, when the acceptance
tests are run, each test (minus Facets) can produce a DOM snapshot. These are
then uploaded to Percy where a visual diff can be done against a baseline.
This document covers how Percy will ultimately be used with the SDK:

https://docs.google.com/document/d/1yui_jHVq1o8Zg2J9iSSGuym842HJ-tUZE7PrnaH5_nk/edit?usp=sharing

As part of this item, I created an answers-sdk Percy organization. In this
organization, I created a v1.4 project and established the visual testing
baselines for the various snapshots.

As part of this PR, I made the 'Universal flow' test a bit simpler. Namely,
I removed the MapComponent for the 'people' vertical. Because the map loads
asynchronously, it was present in some DOM snapshots but not others. This was
leading to Percy spuriously flagging builds as 'changed'.

J=SPR-2433
TEST=auto

Ensured that when I set a PERCY_TOKEN in my bash_profile, Percy snapshots were
generated and uploaded to the v1.4 project. If no token was set, the acceptance
tests ran as before, but not snapshot was generated.