ResultsHeader: hide nlp filters that duplicate another filter (#919)
This commit changes the ResultsHeader component to ignore nlpFilterNodes
that are duplicates of a different applied FilterNode. This way, if
somebody were to search for a nlpFilter like "technology", that is
configured on the back-end to also return a selected facet in the response,
only the facet filter tag will be displayed.

TEST=manual, auto
check that I can search consulting or technology, the facet will be checked
off, and only the removable facet filter tag will be rendered.