Facets/FilterOptions: match old facets behavior with persistent storage
pre v1.4.0 version of the SDK had Facets behavior where FilterOptions
would store checked off options in the url, and on page load these
previously stored options would be used to check off options, but
would not prohibit any options marked as selected from the backend
from being checked.

The Facets component marks options as selected using a selected flag
within FilterOptions' config. As part of v1.4.0 this config was turned
into a public interface, with specification that it would be used to
check off options ONLY if none were stored in the url. This is at
odds with the previous Facets behavior.

This was fixed by adding additional isDynamic logic to FilterOptions
when setting previously selected options.

note: since the this.name of a FilterOptions used in a Facets depends on
it's index in the Facets, and since this order depends on the query
(e.g. a query with "Consulting" might only show 4 Facets, and have the Awards Facet
as the 4th one, while a blank search might have it as the 11th one). This causes
behavior where if you select a facet, run a search that changes the order/length
of facets, then select another facet, since refreshing the page uses the first
set of facets not the second, the previously selected facets will not be
selected on page refresh. This is consistent with previous behavior.

TEST=manual
- test that reported bug is fixed
search for 'Consulting', see that facet is checked off and removable filter tag
is displayed. Uncheck 'Consulting' facet, refresh the page. Previously, the 'Consulting'
facet would no longer be checked, now it is. The removable filter tag is also displayed.

- test that we match 0.12.2 behavior
In a filterbox filtering on c_awards, select 'Roses Friend', apply filters. Select filterbox's
'Pres Club', apply filters. See that Roses Friend and Presidents Club facets are checked off
Refresh page, only see Roses Friend facet is checked off.

Select 'Pres Club' in filterbox, apply filters. Select 'Roses Friend' in filterbox, apply filters.
See that both Roses Friend and Presidents Club facets are checked off.
Refresh page, see that both Roses Friend and Presidents Club facets are checked off.