Update behavior for empty search when redirect url is provided (#953)
Fix behavior of empty search when redirect url is provided. Previously, when a searchBar is initialized with a redirectUrl param, it's possible to submit an empty search. Now, if it's a search bar for a vertical search (verticalKey has been included), we use the allowEmptySearch boolean to control empty search behavior. If it's a search bar for a universal search, an empty search is not be allowed.

J=SPR-2470

TEST=manual

Tested on local site. Tested behavior when search is empty/not empty, verticalKey is included/not included and when allowEmptySearch is true/false.