query rules: Set persistentStorage on search instead of on init or setContext
Similar to the query, we only set referrerPageUrl or context for the
first time in the URL after a search. This is to mitigate changes to
client subdomains. Before, referrerPageUrl would be added to the URL on
ANSWERS.init, even when the client only had a redirectUrl searchbar.

We change it so that after /any/ search to the API, we update the query
params. A search on universal, vertical, pagination, facets, will all
change the query parameters to any updated values. Note: they are still
available on globalStorage. We are moving the persistentStorage updates.

J=SPR-2554
TEST=manual

Test a searchbar with a redirectUrl. When searching, the referrerPageUrl
and the context do not show on the URL, only on the redirected URL.

Test that setContext context is not applied on a regular experience
until search or pagination or facet apply (essentially any re-search).

Test that referrerPageUrl is still added to links in navigation, view
more, alternative verticals, even if it is not currently in the URL.