Removed default params value in getTabUrl and changed corresponding test (#949)
Remove default value for params parameter in getTabUrl in navigationcomponent.js and updated test that checks if tab url can be generated without params. The default value was added to fix a bug where the nav links had empty href's on the landing page, but got populated after submitting a query. We don't need the default value anymore as params in the calls to getTabUrl are given a value.

J=SPR-2518

TEST=manual

Tested on local site by making sure site can build and functions work properly.