UniversalResults: add top level appliedFilters config
This commit adds a top level appliedFilters config
to UniversalResults. This allows users to easily specify
shared config for appliedFilters, and override it
per vertical.

J=SPR-2540
TEST=manual

Tested that each config option specified in
getChildConfig can be set in the global config,
as well as overridden for a specific vertical.