searchbar: re-allow redirectUrl without verticalKey (#963)
If a user wants to add a redirectUrl without a verticalKey, it should
run a universal search with no vertical key. Before this change, we
would not allow this behavior.

J=none
TEST=manual

Tested that you can run a search on a searchbar with a redirect URL and
no vertical key and this runs a universal search on the redirect
experience.