query rules: Set persistentStorage on search instead of on init or setContext (#960)
* query rules: Set persistentStorage on search instead of on init or setContext

Similar to the query, we only set referrerPageUrl or context for the
first time in the URL after a search. This is to mitigate changes to
client subdomains. Before, referrerPageUrl would be added to the URL on
ANSWERS.init, even when the client only had a redirectUrl searchbar.

We change it so that after /any/ search to the API, we update the query
params. A search on universal, vertical, pagination, facets, will all
change the query parameters to any updated values. Note: they are still
available on globalStorage. We are moving the persistentStorage updates.

J=SPR-2554
TEST=manual

Test a searchbar with a redirectUrl. When searching, the referrerPageUrl
and the context do not show on the URL, only on the redirected URL.

Test that setContext context is not applied on a regular experience
until search or pagination or facet apply (essentially any re-search).

Test that referrerPageUrl is still added to links in navigation, view
more, alternative verticals, even if it is not currently in the URL.

* Tests remove the storage key expect, default to null

* Fix back-button for query rules by only updating url when different (#967)

* Fix back-button for query rules by only updating url when different

Our persistent storage updates history in batches. Imagine the following
scenario:

```
persistentStorage.set(QUERY, ..., replaceHistory = false);
persistentStorage.set(REFERRER_PAGE_URL, ..., replaceHistory = false);
persistentStorage.set(CONTEXT, ..., replaceHistory = true);
```

Because of our implementation oof batched updates, all three will be set
at the same time with replaceHistory = true (Even if the first two were
applied with false).

Our implementation of query rules applies the query params to
the url right before a search. This is at the same time query is
applied. Thus, we should keep replaceHistory to match what query is
doing (replaceHistory = false).

This puts in a problem when we land on an experience with a `query`
query param. Because replaceHistory must be false and because we set
context/referrerPageUrl before any search, we push another state
when we land on this page. This makes it so you cannot go backwards in
browser history from this page. (See test example below).

By only updating the URL when query params are different, we do not push
that extra state on page land. This is also useful for avoiding pushing
extra unnecessary state.

J=SPR-2554
TEST=manual

Test that you can make three queries in succession and you are not stuck
in the second query. For example.

Land on empty universal.
Search for virginia
Search for all
Search for christian

If you go back in your browser history twice, you should be back on the
`virginia` search. Not the `all` search.

* remove unnecessary check in equivalentParams