searchbar: set referrerPageUrl on pageload for non-redirectUrl searchbar
Imagine the following situation

* Page A has a redirect URL searchbar to Page B
* User types in a query on Page A and gets redirected to Page B w/ query
* The page is updated to include referrerPageUrl
* Going back in history takes you to Page B without referrerPageUrl
* referrerPageUrl is added again

If you redirected to an answers experience through a redirectUrl
searchbar, you would not be able to go back in history. This is because
we run a search on load when we see a query is in the query parameters.
Every search from the search component will try to add referrerPageUrl
to the URL.

To fix this, we set the referrerPageUrl with replaceHistory true on
searchbar load when we make the initial query search. Persistent storage
will not update history after this because referrerPageUrl is already in
the URL.

J=SPR-2554
TEST=manual

Test on a local site that if you redirect to an answers experience from
redirectUrl searchbar and then try to go back in history, you are not
stuck.

Test on a local site that if you redirect to an answers experience with
a query from a link and then try to go back in history, you are not
stuck.