core/search: dont replace history for vertical search query rules params
Make the vertical search match the universal search and do not replace
history when setting referrerPageUrl and context. We do this because
otherwise every search would replace history, when we do want to push to
history when there is a new query. This is a result of batching
history changes. The last replaceHistory boolean will be the setting for
all updates to history in the batch.

J=SPR-2554
TEST=manual

Make a few searches on a vertical page, make sure you can go back in
history to see previous queries